requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==10.0.0
flask==2.3.3
werkzeug==2.3.7